Seems to work

+++ common/wave.h
@@ -14,12 +14,12 @@ extern "C"
 
   // Load signal in floating point format (-1 .. +1) as a WAVE file using 16-bit signed integers.
   // Now mallocs signal array, places in *signal, caller must free
-  int load_wav(float** signal, int* num_samples, int* sample_rate, const char* path,int fd);
+  int load_wav(float** signal, int* num_samples, double* sample_rate, const char* path,int fd);
 
   // base_freq = radio frequency in Hz corresponding to zero frequency here (receiver is always USB)
   // tmp = UTC @ signal[0]
   // fsec = fractional second in UTC @ signal[0]
-  int process_buffer(float const *signal,int sample_rate, int num_samples, bool is_ft8, float base_freq, struct tm const *tmp, double fsec);
+  int process_buffer(float const *signal,double sample_rate, int num_samples, bool is_ft8, double base_freq, struct tm const *tmp, double fsec);
 
 #ifdef __cplusplus
 }
